rename: endpoints -> clients
