+++ src/clients/pagination/stream.rs
